Add files via upload
Kevin Hu, CEO of a multichannel apparel and home goods retailer in the US (hereafter called the firm),
looked at the deserted flagship store of the firm on 5th Avenue in NY City. The firm has suffered massive
losses due to the closure of its stores during COVID-19 pandemic. Kevin has to decide now whether to
close stores permanently or wait till the economy opens up. The COO of the company warned Kevin that
their online sales are highly interdependent with their stores. He argued that the firm acquires loyal
customers through stores, and these customers create word of mouth publicity for their brand. By closing
their stores permanently, they risk losing their customers to the competitors. Kevin also recalled the demand
of Angela Perry, representative of the firm’s employee union. Angela resented that store employees are not
rewarded for the firm’s online sales that originate due to customers’ store visits. She argued that sales
associates help customers find desired products on the firm’s website if they cannot find them in the store.
Therefore, a substantial portion of the firm’s online sales is attributable to the stores.
Kevin put together a team of marketing executives and data analysts to seek an answer to the general
question– what is the effect of the firm’s stores on its online sales?